requirements.txt
numpy~=2.4.1
pillow~=12.0.0
piexif~=1.1.3
pyyaml~=6.0.3
loguru~=0.7.3
scipy~=1.16.3
opencv